fix(tests): tests should run in the runner now
